[DX] Move ProductValueKeyGenerator from catalog bundle to component
